requirements.txt
python-telegram-bot
python-dotenv
httpx
pip install "python-telegram-bot[job-queue]"
cachetools